requirements.txt
# SL-PAM and PALM algorithm applied to joint image restauration and edges detection
 
# Requirements:


# pillow>=8.0.1
Pillow == 9.2.0
numpy
matplotlib
scipy


# open-cv
#pip install opencv-python
opencv-python
# skimage
#python -m pip install -U scikit-image
scikit-image

PyQt5